+++ src/main/java/com/sparta/hotitemcollector/domain/product/dto/ProductRequestDto.java
@@ -1,5 +1,6 @@
-package com.sparta.hotitemcollector.domain.product;
+package com.sparta.hotitemcollector.domain.product.dto;
 
+import com.sparta.hotitemcollector.domain.product.entity.ProductCategory;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
