+++ src/main/java/com/sparta/hotitemcollector/domain/product/service/SearchService.java
@@ -1,6 +1,9 @@
-package com.sparta.hotitemcollector.domain.product;
+package com.sparta.hotitemcollector.domain.product.service;
 
-import com.sparta.hotitemcollector.domain.user.User;
+import com.sparta.hotitemcollector.domain.product.repository.ProductRepository;
+import com.sparta.hotitemcollector.domain.product.dto.ProductSimpleResponseDto;
+import com.sparta.hotitemcollector.domain.product.entity.Product;
+import com.sparta.hotitemcollector.domain.product.entity.ProductCategory;
 import com.sparta.hotitemcollector.domain.user.UserService;
 import java.util.List;
 import java.util.stream.Collectors;
